requirements.blackwell.txt
numpy==1.26.2
optuna==4.5.0
pillow==11.3.0
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.15.3
torchmetrics==1.8.2
transformers==4.51.3
wandb==0.22.1
tqdm==4.67.1
